+++ database/migrations/2020_11_18_222549_create_pizza_prices_table.php
@@ -18,7 +18,7 @@ class CreatePizzaPricesTable extends Migration
             $table->timestamp('date')->nullable();
             $table->float('price');
             $table->timestamps();
-            $table->unsignedInteger('pizza_id');
+            $table->unsignedBigInteger('pizza_id');
 
             $table->foreign('pizza_id')->references('id')->on('pizzas')
                   ->onDelete('cascade')->onUpdate('cascade');
